first (kind of) working version \o/

+++ inc/rfc3161.h
@@ -30,7 +30,7 @@ static int reply_command(CONF *conf, char *section, char *engine, char *query,
                          int text);
 static TS_RESP *read_PKCS7(BIO *in_bio);
 int create_response(rfc3161_context *ct, char *query, int query_len,
-                    TS_RESP_CTX *resp_ctx, int *resp_size,
+                    TS_RESP_CTX *resp_ctx, size_t *resp_size,
                     unsigned char **resp);
 static ASN1_INTEGER *serial_cb(TS_RESP_CTX *ctx, void *data);
 static ASN1_INTEGER *next_serial(const char *serialfile);
